Shuffle node into its own file
